+++ jaxley/__init__.py
@@ -13,4 +13,3 @@ from jaxley.io.swc import read_swc
 from jaxley.modules import *
 from jaxley.optimize import ParamTransform
 from jaxley.stimulus import datapoint_to_step_currents, step_current
-from jaxley.utils.misc_utils import deprecated, deprecated_kwargs