+++ jaxley/pumps/__init__.py
@@ -1,6 +1,7 @@
 # This file is part of Jaxley, a differentiable neuroscience simulator. Jaxley is
 # licensed under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
+from jaxley.pumps.pump import Pump  # isort: skip
 from jaxley.pumps.ca_pump import CaPump
-from jaxley.pumps.current_to_concentration_change import CaCurrentToConcentrationChange
-from jaxley.pumps.pump import Pump
+from jaxley.pumps.faraday_electrolysis import CaFaradayConcentrationChange
+from jaxley.pumps.nernstreversal import CaNernstReversal