+++ tests/jaxley_identical/test_basic_modules.py
@@ -15,10 +15,12 @@ from math import pi
 import jax.numpy as jnp
 import numpy as np
 import pytest
+from jaxley_mech.channels.l5pc import CaHVA
 
 import jaxley as jx
 from jaxley.channels import HH
 from jaxley.connect import connect, fully_connect
+from jaxley.pumps import CaFaradayConcentrationChange, CaNernstReversal
 from jaxley.synapses import IonotropicSynapse, TestSynapse
 
 
@@ -195,8 +197,8 @@ def test_cell(voltage_solver, SimpleCell):
     assert max_error <= tolerance, f"Error is {max_error} > {tolerance}"
 
 
-def test_cell_unequal_compartment_number(SimpleBranch):
-    """Tests a cell where every branch has a different number of compartments."""
+def test_complex_cell(SimpleBranch):
+    """Test cell with a variety of channels, pumps, diffusion, and comps per branch."""
     dt = 0.025  # ms
     current = jx.step_current(
         i_delay=0.5, i_dur=1.0, i_amp=0.1, delta_t=0.025, t_max=5.0
@@ -209,24 +211,52 @@ def test_cell_unequal_compartment_number(SimpleBranch):
     cell = jx.Cell([branch1, branch2, branch3, branch4], parents=[-1, 0, 0, 1])
     cell.set("axial_resistivity", 10_000.0)
     cell.insert(HH())
-    cell.branch(1).comp(1).stimulate(current)
-    cell.branch(0).comp(0).record()
-    cell.branch(2).comp(2).record()
-    cell.branch(3).comp(1).record()
-    cell.branch(3).comp(3).record()
 
-    voltages = jx.integrate(cell, delta_t=dt, voltage_solver="jax.sparse")
-    voltages_170924 = jnp.asarray(
+    # Calcium mechanisms.
+    cell.branch(0).insert(CaHVA())
+    cell.branch(0).set("CaHVA_gCaHVA", 0.0001)  # Larger than default to see impact on v
+    cell.insert(CaFaradayConcentrationChange())
+    cell.insert(CaNernstReversal())
+    cell.diffuse("CaCon_i")
+    # Large value to see the diffusion after a few milliseconds already.
+    cell.set("axial_diffusion_CaCon_i", 180.2)
+
+    cell.branch(1).comp(1).stimulate(current)
+    cell.branch(0).comp(0).record("v")
+    cell.branch(2).comp(2).record("v")
+    cell.branch(3).comp(1).record("v")
+    cell.branch(3).comp(3).record("v")
+    cell.branch(0).comp(0).record("CaCon_i")
+    cell.branch(2).comp(2).record("CaCon_i")
+    cell.branch(3).comp(1).record("CaCon_i")
+    cell.branch(3).comp(3).record("CaCon_i")
+
+    recordings = jx.integrate(cell, delta_t=dt, voltage_solver="jax.sparse")
+    voltages_240225 = jnp.asarray(
         [
-            [-70.0, -53.80615874, 22.05497283, -47.39519176, -75.64644816],
-            [-70.0, -61.74975007, 34.11362303, -28.88940829, -75.73717223],
-            [-70.0, -52.99972332, 20.97523461, -47.83840746, -75.66974145],
-            [-70.0, -60.86509139, 35.1106244, -33.89684951, -75.76946014],
+            [-70.0, -53.80615874, 22.05557323, -47.39215035, -75.64505091],
+            [-70.0, -61.74975007, 34.11362887, -28.88938285, -75.73709452],
+            [-70.0, -52.99972332, 20.97523502, -47.8383916, -75.66971367],
+            [-70.0, -60.86509139, 35.11062442, -33.89684872, -75.76945538],
         ]
     )
-    max_error = np.max(np.abs(voltages[:, ::50] - voltages_170924))
-    tolerance = 1e-8
-    assert max_error <= tolerance, f"Error is {max_error} > {tolerance}"
+    cacon_240225 = (
+        jnp.asarray(
+            [
+                [5.0, 5.0, 5.02947566, 5.1846641, 5.21696816],
+                [5.0, 5.0, 5.00021622, 5.01602762, 5.06546814],
+                [5.0, 5.0, 5.00003089, 5.00430336, 5.0223509],
+                [5.0, 5.0, 5.00000078, 5.00059561, 5.00648883],
+            ]
+        )
+        * 1e-5
+    )
+    max_error_ca = np.max(np.abs(recordings[4:, ::50] - cacon_240225))
+    max_error_v = np.max(np.abs(recordings[:4, ::50] - voltages_240225))
+    tolerance_ca = 1e-10
+    tolerance_v = 1e-8
+    assert max_error_ca <= tolerance_ca, f"Ca Error is {max_error_ca} > {tolerance_ca}"
+    assert max_error_v <= tolerance_v, f"Voltage Error is {max_error_v} > {tolerance_v}"
 
 
 @pytest.mark.parametrize("voltage_solver", ["jaxley.stone", "jax.sparse"])